+++ system/system.go
@@ -1,7 +1,6 @@
 package system
 
 import (
-	"github.com/hajimehoshi/ebiten"
 	"github.com/kyeett/ecs/events"
 )
 
@@ -9,9 +8,4 @@ import (
 type System interface {
 	Send(events.Event)
 	Update(diff float64)
-}
-
-// A RenderSystem is updated every iteration, and draws to a screen
-type RenderSystem interface {
-	Update(*ebiten.Image)
 }
